requirements.txt
art==6.5
requests==2.32.5
discord.py-self==2.1.0